Adding Guard in the task controller and adding the user in all the operations of the tasks (CRUD)

+++ src/tasks/tasks.service.ts
@@ -8,6 +8,7 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { Task } from './task.entity';
 // import { TaskStatus } from './tasks.model';
 import { TaskStatus } from './task-status.enum';
+import { User } from '../auth/user.entity';
 
 @Injectable()
 export class TasksService {
@@ -17,8 +18,8 @@ export class TasksService {
     private taskRepository: TaskRepository) {
   }
 
-  public async getTaskById(id: number): Promise<Task> {
-    const foundTask = await this.taskRepository.findOne(id);
+  public async getTaskById(id: number, user: User): Promise<Task> {
+    const foundTask = await this.taskRepository.findOne({ where: { id, userId: user.id }});
 
     if (!foundTask) {
       throw new NotFoundException(`Task with <<${id}>> not found`);
@@ -27,7 +28,7 @@ export class TasksService {
     return foundTask;
   }
 
-  public async createTask(createTaskDTO: CreateTaskDTO): Promise<Task> {
+  public async createTask(createTaskDTO: CreateTaskDTO, user: User): Promise<Task> {
     // NOTE: This code commented is the implementation of the method using the ENTITY directly here
     // const { title, description } = createTaskDTO;
     // const task = new Task();
@@ -38,11 +39,11 @@ export class TasksService {
     // await task.save();
     // return task;
 
-    return this.taskRepository.createTask(createTaskDTO);
+    return this.taskRepository.createTask(createTaskDTO, user);
   }
 
-  public async deleteTaskById(id: number): Promise<void> {
-    const result = await this.taskRepository.delete(id); //Here is not necessary to call logic in the repository because we can call the delete right here
+  public async deleteTaskById(id: number, user: User): Promise<void> {
+    const result = await this.taskRepository.delete({ id, userId: user.id }); //Here is not necessary to call logic in the repository because we can call the delete right here
 
     if (result.affected === 0) {
       throw new NotFoundException(`Task with <<${id}>> not found`);
@@ -51,16 +52,16 @@ export class TasksService {
     return;
   }
 
-  public async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-    const taskFound = await this.getTaskById(id);
+  public async updateTaskStatus(id: number, status: TaskStatus, user: User): Promise<Task> {
+    const taskFound = await this.getTaskById(id, user);
     taskFound.status = status;
     await taskFound.save();
 
     return taskFound;
   }
 
-  public async getTasks(filterDTO: GetTasksFilterDTO): Promise<Task[]> {
-    return this.taskRepository.getTasks(filterDTO);
+  public async getTasks(filterDTO: GetTasksFilterDTO, user: User): Promise<Task[]> {
+    return this.taskRepository.getTasks(filterDTO, user);
   }
 
   /* NOTE: REMOVE THE COMMENT IN THIS SECTION IN ORDER TO RUN THE APPLICATION WITHOUT DATABASE AND ENTITIES NOR TYPEORM
